LayoutTests/fast/js/date-big-setdate-expected.txt

This test checks for regression against: 3381 Date.prototype.setDate() incorrect for values >=128.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS d.valueOf() - curValue is millisecondsPerDay
PASS d.valueOf() - curValue is millisecondsPerDay
PASS d.valueOf() - curValue is millisecondsPerDay
PASS d.valueOf() - curValue is millisecondsPerDay
PASS d.valueOf() - curValue is millisecondsPerDay
FAIL d.valueOf() - curValue should be 86400000 (of type number). Was 82800000 (of type number).
PASS d.valueOf() - curValue is millisecondsPerDay
PASS d.valueOf() - curValue is millisecondsPerDay
PASS d.valueOf() - curValue is millisecondsPerDay
PASS successfullyParsed is true

TEST COMPLETE
